requirements.txt
python-telegram-bot
google-genai
flask
requests